refactor: extract game logic function
소켓 커스텀 이벤트에서 사용하는 함수들 추출
